image_classification_api/classifier_api/requirements.txt
flask
flasgger
tensorflow
pillow
mod_wsgi
